+++ src/application/data-access-scope.ts
