+++ src/handlers/AccessApprovalHandler.ts
@@ -1,9 +1,9 @@
 import { GrantScope } from "solid-interoperability";
 import { IAccessApproval } from "../interfaces/IAccessApproval";
-import { Approval } from "../RDF/application/approval";
-import { ApplicationProfileDocument } from "../RDF/application/application-profile-document";
-import { AccessNeedGroup } from "../RDF/application/access-need-group";
-import { DataAccessScope } from "../RDF/application/data-access-scope";
+import { Approval } from "../application/approval";
+import { ApplicationProfileDocument } from "../profile-documents/application-profile-document";
+import { AccessNeedGroup } from "../application/access-need-group";
+import { DataAccessScope } from "../application/data-access-scope";
 
 export class AccessApprovalHandler implements IAccessApproval {
   getAccessScope(): GrantScope {
