+++ src/namespace.ts
@@ -1,1 +1,5 @@
+import N3 from "n3";
+const { namedNode } = N3.DataFactory;
+
 export const INTEROP = "https://www.w3.org/ns/solid/interop#";
+export const type_a = namedNode("http://www.w3.org/1999/02/22-rdf-syntax-ns#type");
