+++ src/profile-documents/social-agent-profile-document.ts
@@ -1,7 +1,7 @@
 import N3, { NamedNode, Prefixes } from "n3";
 import { Session } from "@inrupt/solid-client-authn-node";
-import { parseTurtle } from "./utils/turtle-parser";
-import { serializeTurtle } from "./utils/turtle-serializer";
+import { parseTurtle } from "../utils/turtle-parser";
+import { serializeTurtle } from "../utils/turtle-serializer";
 import { DatasetCore } from "@rdfjs/types";
 
 const { Store, Parser, DataFactory } = N3;
@@ -9,18 +9,18 @@ const { quad, namedNode, defaultGraph } = DataFactory;
 
 const oidcIssuer_PREDICATE = "http://www.w3.org/ns/solid/terms#oidcIssuer";
 
-export class ProfileDocument {
+export class SocialAgentProfileDocument {
   constructor(
     public webId: string,
     public dataset: DatasetCore,
     public prefixes: Prefixes,
   ) {}
 
-  static async getProfileDocument(webId: string): Promise<ProfileDocument> {
+  static async getProfileDocument(webId: string): Promise<SocialAgentProfileDocument> {
     let result = await fetch(webId)
       .then((res) => res.text())
       .then((res) => parseTurtle(res, webId));
-    return new ProfileDocument(webId, result.dataset, result.prefixes);
+    return new SocialAgentProfileDocument(webId, result.dataset, result.prefixes);
   }
 
   hasAuthorizationAgent(authorization_uri: string): boolean {
